Modify review counting to favor users who comment on PRs
Currently review counting is based on someone who creates a review or
who someone who is requested as a reviewer. While someone may have been
requested as a reviewer, it does not mean they actually reviewed the PR.

Additionally, someone may comment on a PR without officially making a
review.

Modify the counting mechanism so that anyone who makes a comment or a
review on a PR other than their own is counted as a reviewer.